Use priceIntentArom in PriceIntentWarningDialog
